Add LXMF announce handlers for delivery and propagation aspects
