+++ src/router/mod.rs
@@ -1,7 +1,10 @@
+mod handlers;
 mod router;
+pub use handlers::{
+    LXMFDeliveryAnnounceHandler, LXMFPropagationAnnounceHandler, PropagationNodeAnnounceData,
+};
 pub use router::{
-    APP_NAME, DELIVERY_ASPECT, LXMFDeliveryAnnounceHandler, LXMFPropagationAnnounceHandler,
-    LxmRouter, PROPAGATION_ASPECT, PropagationNodeAnnounceData, RouterConfig, STAMP_COST_EXPIRY_S,
+    APP_NAME, DELIVERY_ASPECT, LxmRouter, PROPAGATION_ASPECT, RouterConfig, STAMP_COST_EXPIRY_S,
     display_name_from_app_data, pn_announce_data_is_valid, pn_name_from_app_data,
     stamp_cost_from_app_data,
 };
